meshes ect..

+++ src/core/colors.cpp
@@ -9,7 +9,7 @@ namespace bolt
     }
 
     RGB::RGB(uint8_t r, uint8_t g, uint8_t b, uint8_t a)
-        :r(r), g(g), b(b), a(a), r_dec(r/255.0f), g_dec(g/255.0f), b_dec(b/255.0f), a_dec(a/255.0f)
+        :r(r), g(g), b(b), a(a), r_dec((float)r/255.0f), g_dec((float)g/255.0f), b_dec((float)b/255.0f), a_dec((float)a/255.0f)
     {
         ;
     }
